+++ rosie/__main__.py
@@ -1,0 +1,4 @@
+from rosie import main
+
+
+main()
