+++ tests/test_rosie.py
@@ -1,12 +1,12 @@
 import os.path
 from pathlib import Path
 from tempfile import mkdtemp
-from unittest import TestCase, main
-from unittest.mock import MagicMock, Mock, patch
+from unittest import TestCase
+from unittest.mock import MagicMock, patch
 
-import numpy as np
 import pandas as pd
-from rosie.rosie import Rosie
+
+from rosie import Rosie
 
 
 class TestRosie(TestCase):
@@ -22,18 +22,13 @@ class TestRosie(TestCase):
         self.dataset = pd.DataFrame().append(row, ignore_index=True)
         self.subject = Rosie(self.dataset, mkdtemp())
 
-    @patch('rosie.rosie.joblib')
+    @patch('rosie.joblib')
     def test_load_trained_model_trains_model_when_not_persisted(self, _):
         model = self.subject.load_trained_model(MagicMock)
         model.fit.assert_called_once_with(self.dataset)
 
-    @patch('rosie.rosie.joblib')
+    @patch('rosie.joblib')
     def test_load_trained_model_doesnt_train_model_when_already_persisted(self, _):
         Path(os.path.join(self.subject.data_path, 'magicmock.pkl')).touch()
         model = self.subject.load_trained_model(MagicMock)
         model.fit.assert_not_called()
-
-
-
-if __name__ == '__main__':
-    main()
